requirements.txt
deepforest
numpy
pandas
opencv-python
pillow
setuptools==66.1.1
utm
keplergl